requirements.txt
asgiref==3.5.0
colorama==0.4.4
Django==4.0.2
django-bootstrap-datepicker-plus==4.0.0
django-cities==0.6
django-credit-cards==0.4.1
django-model-utils==4.2.0
django-static-jquery3==5.1.0
django-yearmonth-widget==0.2.1
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
tqdm==4.63.1
tzdata==2021.5
whitenoise==6.0.0
